refactor: Player 필드를 private화
- 자식들이 부모 필드를 상속하여 호출하는 것이 아닌, 부모 public 메서드로 메세지를 보내도록 변경
