+++ src/main/java/blackjack/domain/player/Dealer.java
@@ -3,7 +3,6 @@ package blackjack.domain.player;
 public class Dealer extends Player {
 
     private static final int DEALER_GET_CARD_UPPER_BOUND = 17;
-
     private static final String NAME = "딜러";
 
     public Dealer() {
@@ -12,7 +11,7 @@ public class Dealer extends Player {
 
     @Override
     public boolean isNotFinished() {
-        return playingCards.getCardSum() < DEALER_GET_CARD_UPPER_BOUND;
+        return getSumOfCards() < DEALER_GET_CARD_UPPER_BOUND;
     }
 
     @Override
