+++ src/test/java/blackjack/domain/player/PlayerTest.java
@@ -1,10 +1,7 @@
-package blackjack.domain;
+package blackjack.domain.player;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import blackjack.domain.player.Dealer;
-import blackjack.domain.player.Gambler;
-import blackjack.domain.player.Player;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
